--- src/main/kotlin/com/marsrover/NasaMarsRover/rover/adapter/in/web/presentation/RoverController.kt
@@ -1,25 +1,0 @@
-package com.marsrover.NasaMarsRover.rover.adapter.`in`.web.presentation
-
-//import com.marsrover.NasaMarsRover.presentation.view.renderPlateauView
-//import com.marsrover.NasaMarsRover.presentation.view.renderStartPage
-//import org.springframework.http.ResponseEntity
-//import org.springframework.stereotype.Controller
-//import org.springframework.ui.Model
-//import org.springframework.web.bind.annotation.GetMapping
-//import org.springframework.web.bind.annotation.PostMapping
-//import org.springframework.web.bind.annotation.RequestBody
-//import org.springframework.web.bind.annotation.RestController
-//
-//@Controller
-//class RoverController() {
-//
-//    @GetMapping("/rovers")
-//    fun createRovers(): ResponseEntity<String> {
-//        // At this point, `roverRequest.numberOfRovers` will contain the number entered in the request
-//        // You can use this variable as needed
-//
-//        val message = "Rovers amount is 1"
-//        return ResponseEntity.ok(message)
-//    }
-//
-//}
